test: move some test code and add location test
