+++ src/fhir/patient.rs
@@ -643,7 +643,7 @@ fn field_extension(url: String, ext_value: ExtensionValue) -> Result<FieldExtens
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::config::{FallConfig, Fhir, LocationConfig, PatientConfig};
+    use crate::fhir::sharded_fhir_functions::get_test_config;
     use fhir_model::Date;
     use fhir_model::DateTime;
     use fhir_model::r4b::codes::HTTPVerb::Delete;
@@ -725,7 +725,7 @@ PID|1|9999999|9999999|88888888|Nachname^SäuglingVorname^^^^^L||202511022120|M||
 
     #[test]
     fn test_create_patient_merge() {
-        let config = test_config();
+        let config = &get_test_config();
 
         let msg =
             Message::parse_with_lenient_newlines(r#"MSH|^~\&|ORBIS|KH|WEBEPA|KH|20230912105234||ADT^A40^ADT_A39|12345678|P|2.5||123456789|NE|NE||8859/1
@@ -788,7 +788,7 @@ MRG|09876543|||09876543|||Musterfrau^Maxi^^^^^L"#, true)
     }
     #[test]
     fn test_delete_patient() {
-        let config = test_config();
+        let config = &get_test_config();
 
         let msg = Message::parse_with_lenient_newlines(r#"MSH|^~\&|ORBIS|KH|WEBEPA|KH|20221121142711||ADT^A29^ADT_A21|71546182|P|2.5||684450133|NE|NE||8859/1
 EVN|A29|202211211427||12127_684450133|MEDCO-TOBL|202211211427
@@ -813,28 +813,13 @@ PID|1|1234567|1234567||Test-UCH^Endoprothese^^^^^L~Test^^^^^^B||19450201|M|||Bal
         );
     }
 
-    fn test_config() -> Fhir {
-        Fhir {
-            facility_id: "260620431".to_string(),
-            meta_source: "test".to_string(),
-            person: PatientConfig {
-                profile: Default::default(),
-                system: "https://fhir.diz.uni-marburg.de/sid/patient-id".to_string(),
-                other_insurance_system:
-                    "https://fhir.diz.uni-marburg.de/sid/patient-other-insurance-id".to_string(),
-            },
-            fall: FallConfig::default(),
-            location: LocationConfig::default(),
-        }
-    }
-
     #[test]
     fn test_map_versicherung_missing_insurance_number() {
         let msg = Message::parse_with_lenient_newlines(r#"MSH|^~\&|ORBIS||RECAPP|ORBIS|201111280725||ADT^A04|11657277|P|2.5|||||DE||DE
 IN1|1||AOK HSA HESSEN|AOK - Die Gesundheitskasse in Hessen-|Musterstrasse 1^^Musterort^^66666^D||||AOK^1^^^1&gesetzlich||||||50001|Mustermann^Max||19500118|Mustergasse 10^^Musterort^^33333^D|||2|||||||201108220723||R||||||||||||M| ^^^^^D  |||||454874316^^^^^^^20150630"#, true).unwrap();
         let in1 = msg.segment("IN1").unwrap();
 
-        let result = map_versicherungsdaten(in1, &test_config()).unwrap();
+        let result = map_versicherungsdaten(in1, &get_test_config()).unwrap();
 
         // Assert
         assert!(result.is_none());
@@ -850,7 +835,7 @@ PV1|1|O|NEPPOLAMB^^^NEP^NEP^000000|R||||44444ARZT^Arzt^Hans Jürgen^^Praxis^^Dr.
 IN1|1||555555555^^^^NII~22222^^^^NIIP~AOK|AOK - Die Gesundheitskasse in Hessen-|Musterstrasse 1^^Musterort^^66666^D||||AOK^1^^^1&gesetzlich|||20020120|20091231||50001|Mustermann^Max||19500118|Mustergasse 10^^Musterort^^33333^D|||2|||||||201108220723||R|||||A454874316|||||||M| ^^^^^D  |||||A454874316^^^^^^^20150630
 "#, true).unwrap();
 
-        let actual = map_versicherungsdaten(msg.segment("IN1").unwrap(), &test_config())
+        let actual = map_versicherungsdaten(msg.segment("IN1").unwrap(), &get_test_config())
             .unwrap()
             .unwrap();
 
@@ -937,7 +922,7 @@ IN2|4||12345TES^TEST GmbH||||||||||||||||||||||||||^PC^0.0||||DE|||N|||kl|||||||
             true,
         ).unwrap();
 
-        let config = test_config();
+        let config = &get_test_config();
         let identifiers = create_patient_identifiers(&msg, &config).unwrap();
         assert_eq!(identifiers.len(), 2);
     }
@@ -953,7 +938,7 @@ IN1|1||8888888888^^^^NII~P DEMO^^^^XX|AOK Hessen|^^Marburg^^35039^D||||AOK^1^^^1
 IN2|1||||||||||||||||||||||||||||^PC^0^K
 IN1|2|00000001|5555555^^^^NII~P DEMO^^^^XX|AOK - Die Gesundheitskasse in Hessen-|Musterstrasse 1^^Musterort^^66666^D||||AOK^1^^^1&gesetzlich|||20091231|||50001|Mustermann^Max||19500118|Mustergasse 10^^Musterort^^33333^D|||2|||||||||R|||||K454874316|||||||M| ^^^^^D  |||||K454874316^^^^^^^20150630
 IN2|2||R^Rentner||||||||||||||||||||||||||^PC^0^K"#, true).unwrap();
-        let config = test_config();
+        let config = &get_test_config();
         let identifiers = create_patient_identifiers(&msg_full, &config).unwrap();
 
         assert_eq!(identifiers.len(), 2);
@@ -996,7 +981,7 @@ IN1|1||8888888888^^^^NII~P DEMO^^^^XX|AOK Hessen|^^Marburg^^35039^D||||AOK^1^^^1
 IN2|1||||||||||||||||||||||||||||^PC^0^K
 IN1|2|00000001|5555555^^^^NII~P DEMO^^^^XX|AOK - Die Gesundheitskasse in Hessen-|Musterstrasse 1^^Musterort^^66666^D||||AOK^1^^^1&gesetzlich||||20091231||50001|Mustermann^Max||19500118|Mustergasse 10^^Musterort^^33333^D|||2|||||||||R|||||K454874316|||||||M| ^^^^^D  |||||K454874316^^^^^^^20150630
 IN2|2||R^Rentner||||||||||||||||||||||||||^PC^0^K"#, true).unwrap();
-        let config = test_config();
+        let config = &get_test_config();
         let identifiers = create_patient_identifiers(&msg_full, &config).unwrap();
         assert_eq!(identifiers.len(), 2);
 
@@ -1038,7 +1023,7 @@ IN1|1||8888888888^^^^NII~P DEMO^^^^XX|AOK Hessen|^^Marburg^^35039^D||||AOK^1^^^1
 IN2|1||||||||||||||||||||||||||||^PC^0^K
 IN1|2|00000001|5555555^^^^NII~P DEMO^^^^XX|AOK - Die Gesundheitskasse in Hessen-|Musterstrasse 1^^Musterort^^66666^D||||AOK^1^^^1&gesetzlich||||||50001|Mustermann^Max||19500118|Mustergasse 10^^Musterort^^33333^D|||2|||||||||R|||||454874316|||||||M| ^^^^^D  |||||454874316^^^^^^^20150630
 IN2|2||R^Rentner||||||||||||||||||||||||||^PC^0^K"#, true).unwrap();
-        let config = test_config();
+        let config = &get_test_config();
         let identifiers = create_patient_identifiers(&msg_full, &config).unwrap();
 
         assert_eq!(identifiers.len(), 2);
